requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
anthropic>=0.39.0
pdfplumber>=0.11.0
pytesseract>=0.3.10
Pillow>=10.0.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
